refactor: split data.ts god module into 7 focused modules
